requirements.txt
gunicorn
mandrill
mailchimp
facebook-sdk
twython
pillow